* change floats for doubles since float are kind of random with small numbers * improve LCL PDF reader * add getId() to TransactionReaderFactory

+++ budget-transactionreader/src/main/java/org/mortagne/budget/internal/transaction/io/qif/QIFTransactionReader.java
@@ -110,7 +110,7 @@ public class QIFTransactionReader extends AbstractLogEnabled implements Transact
                     transaction.setType(value);
                     break;
                 case 'T':
-                    transaction.setValue(Float.valueOf(value));
+                    transaction.setValue(Double.valueOf(value));
                     break;
                 default:
                     break;
